feat: split out BridgeCard into its own component
